requirements.txt
streamlit>=1.20
requests>=2.28
python-dotenv>=1.0
SQLAlchemy>=1.4
pymysql>=1.0
pandas>=1.5
mysql-connector-python>=8.0
